requirements.txt
numpy
flake8
flake8-functions
ffmpeg
codecov
coverage
flask
PyYAML